# src/docs/swagger.yaml
definitions:
  viewmodels.CustomerCreateViewModel:
    properties:
      contacted:
        type: boolean
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  viewmodels.CustomerEditViewModel:
    properties:
      contacted:
        type: boolean
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  viewmodels.CustomerViewModel:
    properties:
      contacted:
        type: boolean
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
paths:
  /customers:
    get:
      consumes:
      - application/json
      description: get customers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/viewmodels.CustomerViewModel'
            type: array
      summary: Show a list of customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: add by json customer
      parameters:
      - description: Add Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/viewmodels.CustomerCreateViewModel'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created
          schema:
            $ref: '#/definitions/viewmodels.CustomerViewModel'
        "400":
          description: Bad Request
      summary: Create a new customer
      tags:
      - customers
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: delete by customer ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted
        "400":
          description: Bad Request
      summary: Delete a customer
      tags:
      - customers
    get:
      consumes:
      - application/json
      description: get customer by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/viewmodels.CustomerViewModel'
      summary: Show a customer
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: update by json customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/viewmodels.CustomerEditViewModel'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated
          schema:
            $ref: '#/definitions/viewmodels.CustomerViewModel'
        "400":
          description: Bad Request
      summary: Update an existing customer
      tags:
      - customers
swagger: "2.0"